requirements.txt
python-dotenv==0.9.1
requests==2.32.3
sanitize-filename==1.2.0
beautifulsoup4==4.12.3